+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Library] `linear progress indicator` component (Orange-OpenSource/ouds-ios#1509)
 - [Library] `circular progress indicator` component (Orange-OpenSource/ouds-ios#409)
 - [DesignToolbox] Sandbox / debug view for experiments and easy debugging
+- [Library] `static list item` and `navigation list item` components (Orange-OpenSource/ouds-ios#265)
 - [DesignToolbox] Support of `tvOS` in Design System Toolbox app and keep `watchOS` on light app (Orange-OpenSource/ouds-ios#1480)
 - [Library] `Tab view` component to use SwiftUI `Tab` API (Orange-OpenSource/ouds-ios#1452)
 - [Library] `View modifier` to force keyboard closing on tap (Orange-OpenSource/ouds-ios#1530)